moved custom error and warnings classes from utilities to base
