+++ brahmap/utilities/__init__.py
@@ -8,10 +8,6 @@ from importlib.util import find_spec
 from .tools import (
     bash_colors,
     modify_numpy_context,
-    TypeChangeWarning,
-    LowerTypeCastWarning,
-    filter_warnings,
-    ShapeError,
     profile_run,
     output_profile,
 )
@@ -30,10 +26,6 @@ __all__ = __all__ + [
     # tools.py
     "bash_colors",
     "modify_numpy_context",
-    "TypeChangeWarning",
-    "LowerTypeCastWarning",
-    "filter_warnings",
-    "ShapeError",
     "profile_run",
     "output_profile",
 ]
