--- docs/api_reference/cpp_extensions/index.md
@@ -1,3 +1,0 @@
-# C++ Extensions
-
-Under development!
