+++ docs/api_reference/misc/LowerTypeCastWarning.md
@@ -1,0 +1,3 @@
+# `brahmap.base.LowerTypeCastWarning`
+
+::: brahmap.base.LowerTypeCastWarning
