+++ docs/api_reference/misc/ShapeError.md
@@ -1,0 +1,3 @@
+# `brahmap.base.ShapeError`
+
+::: brahmap.base.ShapeError
