+++ docs/api_reference/misc/TypeChangeWarning.md
@@ -1,0 +1,3 @@
+# `brahmap.base.TypeChangeWarning`
+
+::: brahmap.base.TypeChangeWarning
