+++ docs/api_reference/misc/filter_warnings.md
@@ -1,0 +1,3 @@
+# `brahmap.base.filter_warnings`
+
+::: brahmap.base.filter_warnings
