+++ docs/api_reference/misc/index.md
@@ -1,0 +1,6 @@
+# Miscellaneous
+
+- [`TypeChangeWarning`](TypeChangeWarning.md)
+- [`LowerTypeCastWarning`](LowerTypeCastWarning.md)
+- [`filter_warnings`](filter_warnings.md)
+- [`ShapeError`](ShapeError.md)
