--- docs/api_reference/utilities/LowerTypeCastWarning.md
@@ -1,3 +1,0 @@
-# `brahmap.utilities.LowerTypeCastWarning`
-
-::: brahmap.utilities.LowerTypeCastWarning
