--- docs/api_reference/utilities/ShapeError.md
@@ -1,3 +1,0 @@
-# `brahmap.utilities.ShapeError`
-
-::: brahmap.utilities.ShapeError
