--- docs/api_reference/utilities/TypeChangeWarning.md
@@ -1,3 +1,0 @@
-# `brahmap.utilities.TypeChangeWarning`
-
-::: brahmap.utilities.TypeChangeWarning
