+++ docs/api_reference/utilities/index.md
@@ -3,6 +3,3 @@
 ## Tools
 
 - [`modify_numpy_context`](modify_numpy_context.md)
-- [`TypeChangeWarning`](TypeChangeWarning.md)
-- [`LowerTypeCastWarning`](LowerTypeCastWarning.md)
-- [`ShapeError`](ShapeError.md)
