+++ mkdocs.yaml
@@ -132,6 +132,15 @@ nav:
   - API Reference:
     - api_reference/index.md
 
+      
+    - litebird_sim Interface:
+      - api_reference/lbsim/index.md
+      - LBSimProcessTimeSamples: api_reference/lbsim/LBSimProcessTimeSamples.md      
+      - LBSim_InvNoiseCovLO_UnCorr: api_reference/lbsim/LBSim_InvNoiseCovLO_UnCorr.md
+      - LBSimGLSParameters: api_reference/lbsim/LBSimGLSParameters.md
+      - LBSim_compute_GLS_maps: api_reference/lbsim/LBSim_compute_GLS_maps.md
+      - LBSimGLSResult: api_reference/lbsim/LBSimGLSResult.md     
+
     - Core Interface:
       - api_reference/core/index.md
       - SolverType: api_reference/core/SolverType.md
@@ -143,22 +152,15 @@ nav:
       - compute_GLS_maps_from_PTS: api_reference/core/compute_GLS_maps_from_PTS.md
       - compute_GLS_maps: api_reference/core/compute_GLS_maps.md
       - separate_map_vectors: api_reference/core/separate_map_vectors.md
-      - GLSResult: api_reference/core/GLSResult.md
-      
-    - litebird_sim Interface:
-      - api_reference/lbsim/index.md
-      - LBSimProcessTimeSamples: api_reference/lbsim/LBSimProcessTimeSamples.md      
-      - LBSim_InvNoiseCovLO_UnCorr: api_reference/lbsim/LBSim_InvNoiseCovLO_UnCorr.md
-      - LBSimGLSParameters: api_reference/lbsim/LBSimGLSParameters.md
-      - LBSim_compute_GLS_maps: api_reference/lbsim/LBSim_compute_GLS_maps.md
-      - LBSimGLSResult: api_reference/lbsim/LBSimGLSResult.md      
+      - GLSResult: api_reference/core/GLSResult.md 
     
+    - Math Functions:
+      - api_reference/math_functions/index.md
+      - parallel_norm: api_reference/math_functions/parallel_norm.md
+
     - Utilities:
       - api_reference/utilities/index.md
       - modify_numpy_context: api_reference/utilities/modify_numpy_context.md
-      - TypeChangeWarning: api_reference/utilities/TypeChangeWarning.md
-      - LowerTypeCastWarning: api_reference/utilities/LowerTypeCastWarning.md
-      - ShapeError: api_reference/utilities/ShapeError.md
     
     - Base Operators:
       - api_reference/base_operators/index.md
@@ -176,11 +178,13 @@ nav:
       - BlockDiagonalLinearOperator: api_reference/base_operators/BlockDiagonalLinearOperator.md
       - BlockHorizontalLinearOperator: api_reference/base_operators/BlockHorizontalLinearOperator.md
       - BlockVerticalLinearOperator: api_reference/base_operators/BlockVerticalLinearOperator.md
-    
-    - Math Functions:
-      - api_reference/math_functions/index.md
-      - parallel_norm: api_reference/math_functions/parallel_norm.md
-    - C++ Extensions: api_reference/cpp_extensions/index.md
+
+    - Miscellaneous: 
+      - api_reference/misc/index.md
+      - TypeChangeWarning: api_reference/misc/TypeChangeWarning.md
+      - LowerTypeCastWarning: api_reference/misc/LowerTypeCastWarning.md
+      - filter_warnings: api_reference/misc/filter_warnings.md
+      - ShapeError: api_reference/misc/ShapeError.md
   
   - Development:
     - development/index.md
